+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add Logo
 - Add voice recognition
 - Add filter list
+- Add restart wishlist button
 
 ### Changed